ml_service/requirements.txt
flask
flask-cors
tensorflow==2.15.0
keras==2.15.0
opencv-python-headless
numpy
Pillow
gunicorn
